LangSmith/ALL_CODE_PRACTICE/requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.1

# Hugging Face + LangChain
transformers>=4.40.0
huggingface-hub>=0.22.0
langchain-core>=0.2.20
langchain-huggingface>=0.1.0

# LangGraph
langgraph>=0.0.50

# Data validation
pydantic>=2.6.0

# Optional but recommended (tokenization stability)
sentencepiece>=0.1.99
streamlit==0.88.0
langchain_core==0.4.0
langchain_huggingface==0.3.1
huggingface_tokenizer==0.1.1
numpy==1.20.2
scipy==1.7.3
aihub==1.3.0
pytorch==1.9.0
torchvision==0.10.0
scikit-learn==1.0.2
python-dotenv==0.20.0